refactor: extract @clawwork/core package with port-based DI
Move stores, session sync, gateway dispatcher, protocol parsing, and
error classification into a new @clawwork/core package. Desktop stores
become thin re-exports wired through an ElectronPlatformAdapter.

- Port interfaces: PersistencePort, GatewayTransportPort, SettingsPort,
  NotificationsPort
- Factory functions: createMessageStore, createTaskStore, createUiStore,
  createSessionSync, createGatewayDispatcher
- Protocol pure functions extracted from useGatewayDispatcher and
  session-sync (extractText, normalizeAssistantTurns, etc.)
- Main process ws-handlers imports parseToolArgs from core
- Preload types derive IpcResult/ChatAttachment from @clawwork/shared
- Zero consumer file changes (28+ component imports unchanged)
- Viewport-based panel width defaults preserved via getViewportWidth dep
- Dispatcher reset() on unmount preserves remount/HMR lifecycle semantics
