+++ packages/core/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@clawwork/core",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "dependencies": {
+    "@clawwork/shared": "workspace:*",
+    "zustand": "^5.0.11"
+  },
+  "devDependencies": {
+    "typescript": "^5.7.0"
+  }
+}
