+++ packages/core/src/ports/gateway-transport.ts
@@ -1,0 +1,90 @@
+import type { IpcResult } from '@clawwork/shared';
+
+export interface GatewayEvent {
+  event: string;
+  payload: Record<string, unknown>;
+  gatewayId: string;
+  seq?: number;
+}
+
+export interface GatewayStatusEvent {
+  gatewayId: string;
+  connected: boolean;
+  error?: string;
+  serverVersion?: string;
+  reconnectAttempt?: number;
+  maxAttempts?: number;
+  gaveUp?: boolean;
+}
+
+export interface GatewayStatusInfo {
+  connected: boolean;
+  name: string;
+  error?: string;
+  serverVersion?: string;
+}
+
+export interface GatewayStatusMap {
+  [gatewayId: string]: GatewayStatusInfo;
+}
+
+export interface GatewayListItem {
+  id: string;
+  name: string;
+  url: string;
+  token?: string;
+  password?: string;
+  pairingCode?: string;
+  authMode?: 'token' | 'password' | 'pairingCode';
+  isDefault?: boolean;
+  color?: string;
+  connected: boolean;
+}
+
+export interface DiscoveredSession {
+  gatewayId: string;
+  taskId: string;
+  sessionKey: string;
+  title: string;
+  updatedAt: string;
+  agentId: string;
+  model?: string;
+  modelProvider?: string;
+  thinkingLevel?: string;
+  inputTokens?: number;
+  outputTokens?: number;
+  contextTokens?: number;
+  messages: { role: string; content: string; timestamp: string }[];
+}
+
+export interface SyncResult {
+  ok: boolean;
+  discovered?: DiscoveredSession[];
+  error?: string;
+}
+
+export interface ChatAttachment {
+  mimeType: string;
+  fileName: string;
+  content: string;
+}
+
+export interface GatewayTransportPort {
+  sendMessage: (
+    gatewayId: string,
+    sessionKey: string,
+    content: string,
+    attachments?: ChatAttachment[],
+  ) => Promise<IpcResult>;
+  chatHistory: (gatewayId: string, sessionKey: string, limit?: number) => Promise<IpcResult>;
+  abortChat: (gatewayId: string, sessionKey: string) => Promise<IpcResult>;
+  patchSession: (gatewayId: string, sessionKey: string, patch: Record<string, unknown>) => Promise<IpcResult>;
+  gatewayStatus: () => Promise<GatewayStatusMap>;
+  syncSessions: () => Promise<SyncResult>;
+  listGateways: () => Promise<GatewayListItem[]>;
+  listModels: (gatewayId: string) => Promise<IpcResult>;
+  listAgents: (gatewayId: string) => Promise<IpcResult>;
+  getToolsCatalog: (gatewayId: string, agentId?: string) => Promise<IpcResult>;
+  onGatewayEvent: (callback: (data: GatewayEvent) => void) => () => void;
+  onGatewayStatus: (callback: (status: GatewayStatusEvent) => void) => () => void;
+}
