+++ packages/core/src/ports/notifications.ts
@@ -1,0 +1,5 @@
+import type { IpcResult } from '@clawwork/shared';
+
+export interface NotificationsPort {
+  sendNotification: (params: { title: string; body: string; taskId?: string }) => Promise<IpcResult>;
+}
