+++ packages/desktop/package.json
@@ -23,6 +23,7 @@
     "postinstall": "electron-builder install-app-deps"
   },
   "dependencies": {
+    "@clawwork/core": "workspace:*",
     "@clawwork/shared": "workspace:*",
     "@fontsource-variable/inter": "^5.2.8",
     "@fontsource-variable/jetbrains-mono": "^5.2.8",
