+++ packages/desktop/tsconfig.json
@@ -8,5 +8,5 @@
     }
   },
   "include": ["src"],
-  "references": [{ "path": "../shared" }]
+  "references": [{ "path": "../shared" }, { "path": "../core" }]
 }
